reorganized, moved code back to foodtruck from main
