+++ api/database.json
@@ -145,66 +145,5 @@
       "price": 5.2
     }
   ],
-  "purchases": [
-    {
-      "id": 1,
-      "entreeId": 1,
-      "sideId": 3,
-      "vegetableId": 2
-    },
-    {
-      "id": 2,
-      "entreeId": 0,
-      "sideId": 0,
-      "vegetableId": 0
-    },
-    {
-      "id": 3,
-      "entreeId": 0,
-      "sideId": 0,
-      "vegetableId": 0
-    },
-    {
-      "id": 4,
-      "entreeId": 7,
-      "sideId": 0,
-      "vegetableId": 0
-    },
-    {
-      "id": 5,
-      "entreeId": 3,
-      "sideId": 3,
-      "vegetableId": 5
-    },
-    {
-      "id": 6,
-      "entreeId": 10,
-      "sideId": 4,
-      "vegetableId": 9
-    },
-    {
-      "id": 7,
-      "entreeId": 1,
-      "sideId": 6,
-      "vegetableId": 1
-    },
-    {
-      "id": 8,
-      "entreeId": 2,
-      "sideId": 3,
-      "vegetableId": 1
-    },
-    {
-      "id": 9,
-      "entreeId": 2,
-      "sideId": 3,
-      "vegetableId": 1
-    },
-    {
-      "id": 10,
-      "entreeId": 10,
-      "sideId": 4,
-      "vegetableId": 9
-    }
-  ]
+  "purchases": []
 }
